size_t used instead int
Former-commit-id: 0f733934af0f1bc8e0c44991b09080186f304b89

+++ tools/IBundle/src/BundlePackager.cpp
@@ -63,7 +63,7 @@ void BundlePackager::mapNames()
 	id=0;
 	for(auto it=fInputData.ressources.begin(); it!=fInputData.ressources.end(); it++){
 		//Compute the new name with the extension
-		int idExtSuffix = it->first.rfind('.');
+		size_t idExtSuffix = it->first.rfind('.');
 		std::stringstream name;
 		name << "Ressources/" << id <<"."<<it->first.substr(idExtSuffix+1);
 		fNamesMap[it->first] = name.str();
